+++ packages/hasher/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.82.0
+
+### Patch Changes
+
+- @fuel-ts/crypto@0.82.0
+- @fuel-ts/interfaces@0.82.0
+- @fuel-ts/utils@0.82.0
+
 ## 0.81.0
 
 ### Patch Changes